remove toInt method for blur radius

+++ Omega/src/com/saggitt/omega/blur/BlurWallpaperFilter.kt
@@ -28,7 +28,7 @@ class BlurWallpaperFilter(private val context: Context) : WallpaperFilter {
     private var blurRadius = 25
 
     override fun applyPrefs(prefs: OmegaPreferences) {
-        blurRadius = prefs.blurRadius.toInt() / BlurWallpaperProvider.DOWNSAMPLE_FACTOR
+        blurRadius = prefs.blurRadius / BlurWallpaperProvider.DOWNSAMPLE_FACTOR
         blurRadius = blurRadius.coerceAtLeast(1).coerceAtMost(25)
     }
